+++ src/domain/index.ts
@@ -1,3 +1,4 @@
 export * from './calculations'
 export * from './estimation'
+export * from './metadata'
 export * from './factories'